+++ app/src/main/java/es/jrodg85/realestate/Vivienda.java
@@ -19,7 +19,7 @@ public class Vivienda {
 
 	// getter
 
-	public long getIdVivienda() {
+	public int getIdVivienda() {
 		return idVivienda;
 	}
